refactor(#87): inventory 필터 타입을 model로 이동
