+++ src/widgets/inventory/model/types.ts
@@ -1,0 +1,7 @@
+export type InventoryEquippedFilter = "ALL" | "EQUIPPED" | "UNEQUIPPED";
+export type InventorySortFilter = "DEFAULT" | "ACQUIRED_DESC" | "ACQUIRED_ASC";
+
+export interface InventoryDateRange {
+  start: string;
+  end: string;
+}
